fixed local storage bugs and uncecessary reloads
