+++ frontend/package.json
@@ -24,7 +24,7 @@
     "dotenv": "^16.3.1",
     "eslint": "8.49.0",
     "eslint-config-next": "13.5.2",
-    "js-tiktoken": "^1.0.7",
+    "js-tiktoken": "^1.0.8",
     "next": "13.5.3",
     "notistack": "^3.0.1",
     "openai": "^4.16.1",
